api_admin/requirements.txt
fastapi>=0.99.1
uvicorn==0.15.0
sqlalchemy==1.4.23
aiosqlite==0.17.0
python-multipart==0.0.5
pyjwt==2.1.0
bcrypt==3.2.0
pydantic>=1.10.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.19.0
email-validator==1.1.3
requests==2.26.0
aiofiles==0.7.0
greenlet==2.0.2

